openssl_curl_example: Fix build error since 1.5 release
The build error in openssl_curl_example that started in the 1.5 release
came from that libcurl entered the ACAP Native SDK. The libcurl in the
SDK took precedence over the custom built libcurl when linking the
example binary. The libcurl from the SDK caused the build error due to a
missing dependency libssh2 that got removed in the SDK build by a
globbing rule mainly meant to remove libssl.

A simple removal of the libcurl files in the SDK path before starting
the example build solves the issue and works with 1.5 release.

The curl build procedure using automake checks the environment and
automatically depends on some libs that it can find in the SDK sysroot,
like zlib and zstd used for compression of transferred data over HTTP.
The libcurl build dependencies have also been skipped via turned off

Solves AxisCommunications/acap-native-sdk-examples#98
Reference: ECODEVT-303