vapix-sh: Add example that uses shellscript only and not C
This example does the same thing as the VAPIX example written in C, but
uses a POSIX shellscript to implement the application instead. It shows
how simpler applications can be written with just a shellscript but still
bring value, which is good for those who are not C-savvy. (Whereas those
who want to build more advanced applications with VAPIX would probably
opt for the C version to build upon.)

Change-Id: I6f0e874996e69e6319107931bbc93b81ffa40db9
Signed-off-by: Joakim Roubert <joakimr@axis.com>