fix(object-detection-cv25): Fix broken link in README
This reverts the link target to what it was before it was changed in 6aa62c57b650ab2d76af35c007933a54e9c6f70b. `vdo-larod-cv25` does not and has never existed in the repository and, since the context of the link did not change, I assume it should still point to the same example as it used to.